Commiting all local changes which was there

+++ src/main/java/com/entrue/pandamic/view/request/stock/BrowserLocation.java
@@ -7,32 +7,32 @@ import java.io.Serializable;
  */
 public class BrowserLocation implements Serializable {
 
-    private Long latitude;
+    private Double latitude;
 
-    private Long longitude;
+    private Double longitude;
 
     public BrowserLocation() {
 
     }
 
-    public BrowserLocation(Long latitude, Long longitude) {
+    public BrowserLocation(Double latitude, Double longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
 
-    public Long getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Long latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Long getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Long longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 }
